+++ libverilua/src/gen/gen_verilua_env_init.rs
@@ -34,7 +34,7 @@ Self {
     lua_time: Duration::default(),
 
     lua,
-    lua_sim_event: None,
+    lua_sim_event: 0,
     lua_main_step: None,
     lua_posedge_step: None,
     lua_negedge_step: None,
@@ -46,35 +46,5 @@ Self {
     has_next_sim_time_cb: false,
 
     #[cfg(feature = "chunk_task")]
-    lua_sim_event_chunk_1: None,
-    #[cfg(feature = "chunk_task")]
-    lua_sim_event_chunk_2: None,
-    #[cfg(feature = "chunk_task")]
-    lua_sim_event_chunk_3: None,
-    #[cfg(feature = "chunk_task")]
-    lua_sim_event_chunk_4: None,
-    #[cfg(feature = "chunk_task")]
-    lua_sim_event_chunk_5: None,
-    #[cfg(feature = "chunk_task")]
-    lua_sim_event_chunk_6: None,
-    #[cfg(feature = "chunk_task")]
-    lua_sim_event_chunk_7: None,
-    #[cfg(feature = "chunk_task")]
-    lua_sim_event_chunk_8: None,
-    #[cfg(feature = "chunk_task")]
-    lua_sim_event_chunk_9: None,
-    #[cfg(feature = "chunk_task")]
-    lua_sim_event_chunk_10: None,
-    #[cfg(feature = "chunk_task")]
-    lua_sim_event_chunk_11: None,
-    #[cfg(feature = "chunk_task")]
-    lua_sim_event_chunk_12: None,
-    #[cfg(feature = "chunk_task")]
-    lua_sim_event_chunk_13: None,
-    #[cfg(feature = "chunk_task")]
-    lua_sim_event_chunk_14: None,
-    #[cfg(feature = "chunk_task")]
-    lua_sim_event_chunk_15: None,
-    #[cfg(feature = "chunk_task")]
-    lua_sim_event_chunk_16: None,
+    lua_sim_event_chunks: [0; SIM_EVENT_CHUNK_MAX],
 }