+++ libverilua/src/vpi_callback.rs
@@ -429,12 +429,7 @@ unsafe extern "C" fn rd_synch_callback(cb_data: *mut t_cb_data) -> PLI_INT32 {
 
     let result = {
         let _readonly_guard = ReadonlyPhaseGuard::new(env);
-        unsafe {
-            env.lua_sim_event
-                .as_ref()
-                .unwrap_unchecked()
-                .call::<()>(user_data.task_id)
-        }
+        env.call_sim_event(user_data.task_id)
     };
 
     if let Err(e) = result {
@@ -505,16 +500,9 @@ unsafe extern "C" fn rw_synch_callback(cb_data: *mut t_cb_data) -> PLI_INT32 {
         }
     }
 
-    unsafe {
-        if let Err(e) = env
-            .lua_sim_event
-            .as_ref()
-            .unwrap_unchecked()
-            .call::<()>(user_data.task_id)
-        {
-            env.finalize();
-            panic!("{}", e);
-        }
+    if let Err(e) = env.call_sim_event(user_data.task_id) {
+        env.finalize();
+        panic!("{}", e);
     }
 
     0
@@ -551,16 +539,9 @@ unsafe extern "C" fn next_sim_time_callback(cb_data: *mut t_cb_data) -> PLI_INT3
         unsafe { Box::from_raw(cb_data.user_data as *mut NormalCbData) };
     let env = get_verilua_env();
 
-    unsafe {
-        if let Err(e) = env
-            .lua_sim_event
-            .as_ref()
-            .unwrap_unchecked()
-            .call::<()>(user_data.task_id)
-        {
-            env.finalize();
-            panic!("{}", e);
-        }
+    if let Err(e) = env.call_sim_event(user_data.task_id) {
+        env.finalize();
+        panic!("{}", e);
     }
 
     0
@@ -859,16 +840,9 @@ unsafe extern "C" fn edge_callback(cb_data: *mut t_cb_data) -> PLI_INT32 {
         #[cfg(feature = "acc_time")]
         let s = std::time::Instant::now();
 
-        unsafe {
-            if let Err(e) = env
-                .lua_sim_event
-                .as_ref()
-                .unwrap_unchecked()
-                .call::<()>(user_data.task_id)
-            {
-                env.finalize();
-                panic!("{}", e);
-            }
+        if let Err(e) = env.call_sim_event(user_data.task_id) {
+            env.finalize();
+            panic!("{}", e);
         }
 
         #[cfg(feature = "acc_time")]
@@ -975,16 +949,9 @@ unsafe extern "C" fn time_callback_handler(cb_data: *mut t_cb_data) -> PLI_INT32
     #[cfg(feature = "acc_time")]
     let s = std::time::Instant::now();
 
-    unsafe {
-        if let Err(e) = env
-            .lua_sim_event
-            .as_ref()
-            .unwrap_unchecked()
-            .call::<()>(task_id)
-        {
-            env.finalize();
-            panic!("{}", e);
-        }
+    if let Err(e) = env.call_sim_event(task_id) {
+        env.finalize();
+        panic!("{}", e);
     }
 
     #[cfg(feature = "acc_time")]
@@ -1115,16 +1082,9 @@ unsafe extern "C" fn edge_callback_always(cb_data: *mut t_cb_data) -> PLI_INT32 
         #[cfg(feature = "acc_time")]
         let s = std::time::Instant::now();
 
-        unsafe {
-            if let Err(e) = env
-                .lua_sim_event
-                .as_ref()
-                .unwrap_unchecked()
-                .call::<()>(user_data.task_id)
-            {
-                env.finalize();
-                panic!("{}", e);
-            }
+        if let Err(e) = env.call_sim_event(user_data.task_id) {
+            env.finalize();
+            panic!("{}", e);
         }
 
         #[cfg(feature = "acc_time")]